Added the jquert css Added a examples dataset
Change the csv from titanic to events (a flag is available) in routines.js

CSVContainer now accept any archive but only support discrete values to categorized, the continuous values are ignored. The type of columns must be manual indicated

The graphic of population now paint using a category and primary column and the graphic can be filtered by categories.

Added examples graphics