Fix: Bootstrap css, now width of a input group in a inline form is not 100 %, now is out Fix: Now the web socket can reconnect with a new direction.
The chart canvas now fit to its div

Add jQuery function to delete the classes that start with a specific prefix

Now, the web-socket is no a global var instead of a attribute of ReconnectingWebSocket. The funciton sendToServer does not work (Must be fixed)

Adding controls to the view. Select the columns, load url dynamically, and load url button (the velocity slider not work must be fixed)
When a new url is set. The view remove the old charts, disconnect the web-sockets and connect to the new url